refactor for å kunne lage mer hensiktsmessig startsside
