+++ src/App.js
@@ -6,7 +6,7 @@ import "./style/GeografiskSidebar.css";
 import "./style/Kartlag.css";
 import "./style/FargeMenyer.css";
 import Grunnkart from "./Grunnkart/Grunnkart";
-import ForsideInformasjon from "./Informasjon/ForsideInformasjon";
+import ForsideInformasjon from "./Forside/ForsideInformasjon";
 import SettingsContainer from "./SettingsContainer";
 import HamburgerMeny from "./HamburgerMeny/HamburgerMeny";
 import MobileNavigation from "MobileNavigation/MobileNavigation";
