+++ src/Forside/HowTo.js
@@ -1,0 +1,27 @@
+<div className="infoItemContainer">
+  <div className="infoItem">
+    <h3>Slå opp landskap og natursystem</h3>
+    <img
+      src="https://data.artsdatabanken.no/Natur_i_Norge/Landskap/forside_408.jpg"
+      alt="Landskap og natursystem"
+    />
+    <p>
+      Hvis du klikker i søkefeltet kan du velge landskap eller natursystem, og
+      navigere deg videre for å se det markert i kartet. Under hierarki kan du
+      flytte deg dypere og dypere inn i systemet, og fargeindikatorene under
+      natursystem vil vise deg hva som er vist i kartet.
+    </p>
+  </div>
+
+  <div className="infoItem">
+    <h3>Finn informasjon om et område</h3>
+    <img
+      src="https://data.artsdatabanken.no/Natur_i_Norge/Landskap/Landskapsgradient/V%C3%A5tmarkspreg/forside_950.jpg"
+      alt="Finn info"
+    />
+    <p>
+      Ved å klikke på kartet kan du sette en markør, og få opp informasjon om
+      det valgte området.
+    </p>
+  </div>
+</div>;
