+++ src/Forside/Kilder.js
@@ -1,0 +1,59 @@
+<div className="infoItemContainer sourceContainer">
+  <h2>Dataleverandører</h2>
+  <div>
+    <img
+      src="https://data.artsdatabanken.no/Datakilde/Artsdatabanken/forside_408.png"
+      alt="logo for artsdatabanken"
+    />
+    <br />
+    Artsdatabanken
+  </div>
+  <div>
+    <img
+      src="https://data.artsdatabanken.no/Datakilde/Milj%C3%B8direktoratet/forside_408.png"
+      alt="logo for miljødirektoratet"
+    />
+    <br />
+    Miljødirektoratet
+  </div>
+  <div>
+    <img
+      src="https://data.artsdatabanken.no/Datakilde/Kartverket/forside_408.png"
+      alt="logo for kartverket"
+    />
+    <br />
+    Kartverket
+  </div>
+  <div>
+    <img
+      src="https://data.artsdatabanken.no/Datakilde/Norsk_institutt_for_naturforskning/forside_408.png"
+      alt="logo for nina"
+    />
+    <br />
+    NINA
+  </div>
+  <div>
+    <img
+      src="https://data.artsdatabanken.no/Datakilde/Norsk_institutt_for_bio%C3%B8konomi/forside_408.png"
+      alt="logo for nibio"
+    />
+    <br />
+    NIBIO
+  </div>
+  <div>
+    <img
+      src="https://data.artsdatabanken.no/Datakilde/Norges_Geologiske_Unders%C3%B8kelse/forside_408.png"
+      alt="logo for ngu"
+    />
+    <br />
+    NGU
+  </div>
+  <div>
+    <img
+      src="https://data.artsdatabanken.no/Datakilde/Statistisk_sentralbyr%C3%A5/forside_408.png"
+      alt="logo for ssb"
+    />
+    <br />
+    SSB
+  </div>
+</div>;
