+++ src/HamburgerMeny/Innstillinger.js
@@ -1,31 +1,17 @@
 import Menyelement from "./Menyelement";
-import { Layers } from "@material-ui/icons";
 import { ListSubheader, Typography } from "@material-ui/core";
-import { withStyles, withTheme } from "@material-ui/core/styles";
 import { SortByAlpha } from "@material-ui/icons";
-import classNames from "classnames";
 import React from "react";
-
-const styles = {
-  link: { cursor: "pointer" },
-  icon: { fontWeight: 600, color: "rgba(0,0,0,0.7)" },
-  icondisabled: { textDecoration: "line-through", color: "rgba(0,0,0,0.4)" }
-};
 
 const Innstillinger = ({
   classes,
   visKoder,
   sorterPåKode,
-  onNavigate,
   onUpdateSetting
 }) => (
   <>
     <ListSubheader>Innstillinger</ListSubheader>
-    <Menyelement
-      //onClick={() => onNavigate("/bakgrunnskart?vis")}
-      icon={<Layers />}
-      primary="Bakgrunnskart"
-    />
+
     <Menyelement
       onClick={e => {
         e.preventDefault();
@@ -34,14 +20,7 @@ const Innstillinger = ({
       }}
       icon={
         <Typography>
-          <span
-            className={classNames(
-              classes.icon,
-              !visKoder && classes.icondisabled
-            )}
-          >
-            NA
-          </span>
+          <span className={!visKoder && classes.icondisabled}>NA</span>
         </Typography>
       }
       primary="Vis koder i tillegg til navn"
@@ -61,4 +40,4 @@ const Innstillinger = ({
   </>
 );
 
-export default withStyles(styles)(withTheme(Innstillinger));
+export default Innstillinger;
