--- src/style/Informasjon.css
@@ -1,183 +1,0 @@
-div {
-  box-sizing: border-box;
-}
-
-/*Card Look materal ui manually bc less work*/
-
-.top_image img,
-.info_component button,
-.infoBox,
-.infoItem {
-  background: #f5f5f5;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
-  border-radius: 4px;
-}
-
-.info_component {
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.6);
-  position: fixed;
-  top: 0;
-  left: 0;
-  text-align: center;
-  padding: 5%;
-  z-index: 50;
-}
-
-.info_component:hover {
-  cursor: pointer;
-}
-
-.infoBox {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  vertical-align: top;
-  overflow-y: auto;
-}
-
-.container_padding {
-  padding: 50px;
-}
-
-.top_image {
-  position: relative;
-  bottom: 0;
-  left: 0;
-  height: 100%;
-  max-height: 60px;
-  border-top-left-radius: 4px;
-  border-top-right-radius: 4px;
-  width: 100%;
-  background: #f5f5f5
-    url(https://www.artsdatabanken.no/Media/F21301?mode=320x320);
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: 100%;
-  z-index: 0;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
-  text-align: left;
-  padding: 0px;
-}
-
-.top_image img {
-  /*The top image logo (our logo)*/
-  margin-top: 10px;
-  margin-left: 10px;
-  margin-right: auto;
-  width: 150px;
-  background: #fff;
-  padding: 10px;
-  border-radius: 4px;
-}
-
-.ingress {
-  font-weight: 400px;
-  font-size: 14pt;
-  margin-bottom: 50px;
-}
-
-.infoItemContainer {
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.infoItem {
-  display: inline-block;
-  flex: 1;
-  padding: 20px;
-  margin: 10px;
-  min-width: 230px;
-}
-
-.infoItem img {
-  width: 150px;
-  height: 150px;
-  border-radius: 50%;
-}
-
-.info_component button {
-  border-radius: 4px;
-  border: 1px solid lightgrey;
-}
-
-.info_component button:hover {
-  cursor: pointer;
-  opacity: 0.5;
-}
-
-.close_button {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  font-weight: bold;
-  height: 40px;
-  width: 40px;
-}
-
-.mobile_close_button {
-  display: none;
-}
-
-.sourceContainer {
-  position: relative;
-  bottom: 0;
-  padding-bottom: 30px;
-  background: #f5f5f5;
-  border-radius: 4px;
-  width: 90%;
-  margin-left: auto;
-  margin-right: auto;
-  box-sizing: border-box;
-}
-.sourceContainer div {
-  flex: 1;
-}
-.sourceContainer h2 {
-  padding: 10px;
-  width: 100%;
-  padding-bottom: 20px;
-}
-
-.sourceContainer img {
-  height: 50px;
-}
-
-@media only screen and (max-width: 768px) {
-  .info_component {
-    padding: 0px;
-  }
-
-  .container_padding {
-    padding: 15px;
-  }
-
-  .infoBox,
-  .top_image {
-    border-radius: 0%;
-  }
-
-  .close_button {
-    display: none;
-  }
-
-  .mobile_close_button {
-    display: inline-block;
-    font-weight: bold;
-    height: 50px;
-    margin-bottom: 30px;
-    font-weight: bold;
-  }
-
-  .sourceContainer div {
-    padding: 10px;
-  }
-}
-
-@media only screen and (min-width: 1300px) {
-  .infoBox {
-    margin: auto;
-    max-width: 1300px;
-  }
-}
